irqbalance: move numa.c to cputree.c
Numa info also belongs to cpu topology, it is no need to separate two
files, so move numa.c to cputree.c.

Merge two files can make the logical reasonable, and can make code
optimization.
